[#785] /lib specific code living on common moved to /lib
